Refactoring
Move errors to its own file
Remove some debugging code
Refactor the tests to pull out some shared code
